Cleaned up schedule components
