+++ src/components/schedule/index.js
@@ -1,0 +1,16 @@
+import ScheduleListItem from './list'
+import { GEList, GEListItem } from './ge-list'
+import ScheduleTermList from './term-list'
+import ScheduleBackLink from './back-link'
+import CourseList from './course-list'
+import CoursePage from './course-page'
+
+export {
+  ScheduleListItem,
+  ScheduleTermList,
+  GEList,
+  GEListItem,
+  CourseList,
+  CoursePage,
+  ScheduleBackLink,
+}
