+++ src/components/schedule/week-days.js
@@ -1,0 +1,30 @@
+export default [
+  {
+    short: 'MON',
+    day: 'Monday',
+  },
+  {
+    short: 'TUES',
+    day: 'Tuesday',
+  },
+  {
+    short: 'WED',
+    day: 'Wednesday',
+  },
+  {
+    short: 'THURS',
+    day: 'Thursday',
+  },
+  {
+    short: 'FRI',
+    day: 'Friday',
+  },
+  {
+    short: 'SAT',
+    day: 'Saturday',
+  },
+  {
+    short: 'SUN',
+    day: 'Sunday',
+  },
+]
